Adding in lambda changes, and a few other various programmatic changes attempting to update it to more standard conventions.

+++ src/main/java/com/tli/orders/entities/LineItem.java
@@ -1,13 +1,11 @@
 package com.tli.orders.entities;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
 
@@ -40,7 +38,7 @@ public class LineItem implements Serializable{
 	@NotNull
 	@Column(name = "price")
 	@Expose
-	private BigDecimal price;
+	private double price;
 	
 	@NotNull
 	@Column(name = "quantity")
@@ -80,11 +78,11 @@ public class LineItem implements Serializable{
 		this.name = name;
 	}
 
-	public BigDecimal getPrice() {
+	public double getPrice() {
 		return price;
 	}
 
-	public void setPrice(BigDecimal price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
